requirements.txt
streamlit
google-generativeai
faiss-cpu
numpy
toml
pysrt